refactor: extract project-http.js from project.js (PR-06)
Move handleHTTP function (617 lines) and parseJsonBody utility into a
dedicated module. Handles all project HTTP routes: extension bridge,
image serving, file upload, push notifications, skill management, git
status, and info endpoint. project.js reduced from 4,120 to 3,497 lines.
